+++ sercom.h
@@ -15,7 +15,7 @@
 #define PACKET_IDENT (('D' << 24) | ('E' << 16) | ('B' << 8) | 'M')
 
 void sercom_init(void);
-void sercom_send(PacketType_e packet_type, const uint8_t *pBuf, uint8_t size);
+void sercom_send(PacketType_e packet_type, const uint8_t *pBuf, uint16_t size);
 bool sercom_receive(PacketHeader_t *pHdr, uint8_t **pPayload);
 
 #endif
